requirements.txt
Django == 5.0.6
psycopg2-binary == 2.9.9
django-ninja == 1.2.0
django-cors-headers == 4.4.0
pydantic~=2.8.2
django-ninja-extra == 0.21.0
django-admin-argon-dashboard==1.0.16